requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.25.0
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.4
django-3-jet==1.0.8
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-import-export==2.5.0
et-xmlfile==1.1.0
feedparser==6.0.5
gunicorn==20.1.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
MarkupPy==1.14
matplotlib==3.4.2
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-magic==0.4.24
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
threadpoolctl==2.1.0
urllib3==1.26.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
